delete

+++ src/notes/notes-service.js
@@ -55,10 +55,7 @@ const notesService = {
 
 
   deleteNote(db, id, user) {
-  return db
-  .from('vocabulab_notes AS notes')
-  .where({card_id: id, user_id: user})
-  .delete();
+    db.raw(`delete from vocabulab_notes where card_id = 1 and user_id = 3` )
 },
